requirements.txt
discord.py==2.3.2
openai==1.3.0
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3
scikit-learn==1.3.2
schedule==1.2.0
aiohttp==3.9.1
asyncio==3.4.3 